requirements.txt
<<<<<<< HEAD
asgiref==3.2.7
astroid==2.4.2
certifi==2020.4.5.2
chardet==3.0.4
colorama==0.4.3
Django==3.0.7
djangorestframework==3.11.0
get==2019.4.13
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.1.2
post==2019.4.13
public==2019.4.13
pylint==2.5.3
pytz==2020.1
query-string==2019.4.13
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.9
wrapt==1.12.1